Make OctetsFrom and OctetsInto have associated error types. (#4)
This commit allows implementations of OctetsFrom and OctetsInto to choose
their own error type. The primary motivation is to allow types where the
conversion can’t fail to use Infallible and provide variations of the
conversion methods that then simply don’t fail.

Thus, the commit also prefixes the names of the primary methods of the two
traits with try_ and adds new try_-less methods that are only available when
conversion can’t fail.