Introduce Octets trait using GATs and drop OctetsRef. (#12)
This commit changes the traits to use the generic associated types aka GATs.
It drops the OctetsRef trait and all its related weirdness in favor of a
simple Octets trait. This new trait serves the same purpose but is
implemented by the octets sequence type itself rather than a reference to it.