Various Clippy-suggested fixes and some fallout from that. (#52)
This PR implements some changes suggested by Clippy. As a result, it
actually provides an implementation of OctetsFrom for Smallvec (which was
presented but never compiled due to a broken cfg attribute). It also removes
the Borrow<[u8]> impls for Str<_> and StrBuilder<_> due to conflicts with
the Hash impl.