Fix Array::resize_raw() to allow resizing to the length of the array. (#63)
The `new_len` parameter to Array::resize_raw() is the new length of the
contained octets sequence. The documentation for resize_raw() specifies
that it is an error for `new_len` to be larger than the array size,
but the current code also returns an error if `new_len` is equal to the
array size.